Move RuntimeErrorState
